Create Inheritance
When an inherited class derives from two base classes, the constructors of the base classes are called first, followed by the constructor of the derived class. However, the destructor of the derived class is called first, followed by the destructor of the two base classes. But if the two base classes have a variable with the same name, it does not matter in which the order the base classes are inherited, it will always give an ambiguity error.